+++ CHANGELOG.md
@@ -35,6 +35,8 @@ Main (unreleased)
   - `phlare.scrape` collects application performance profiles. (@cyriltovena)
   - `phlare.write` sends application performance profiles to Grafana Phlare. (@cyriltovena)
 
+  - `otelcol.receiver.zipkin` receives Zipkin-formatted traces. (@rfratto)
+
 ### Enhancements
 
 - Handle faro-web-sdk `View` meta in app_agent_receiver. (@rlankfo)